Remove chat highlight feature from core
It's too specific to the GUI (only used for window title and tray icon).
